runtime: preserve Win32 function metadata and stacks
