+++ runtime/internal/lib/runtime/funcinfo_entry_slack_windows_386.go
@@ -1,0 +1,13 @@
+//go:build windows && 386
+
+package runtime
+
+// Win32 has no PE unwind table that can validate a guessed function boundary.
+// The compiler therefore emits exact entry PCs for address-taken functions;
+// accepting a post-prologue look-ahead here could misclassify an ordinary PC
+// near the end of one function as the next function.
+const runtimeFuncPCEntrySlack = 0
+
+func runtimeFuncPCMayUseEntrySlack(uintptr) bool {
+	return false
+}
